I added a 'pages' folder where I moved in the 'login.js' and the 'signup.js'. I added routes to the 'index.js' in the client side.
