+++ client/src/App.css
@@ -1,1 +1,3 @@
-
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
