+++ client/src/App.js
@@ -1,22 +1,23 @@
 
 import React from 'react';
 
-import Login from './components/Login';
-import Signup from './components/Signup'
+import Login from './pages/Login';
+import Signup from './pages/Signup'
 import SignupButton from './components/parts/SignupButton';
 import LoginButton from './components/parts/LoginButton';
 
 
-
-
 function App() {
   return (
-    
-      <div>
+
+       <div>
+  
           <Login />
           <Signup />
        </div>
 
+
+     
   );
 }
