+++ client/src/components/parts/LoginButton.js
@@ -1,6 +1,5 @@
 import React from "react";
-import Login from '../Login';
-import { Link } from 'react-router-dom';
+import Login from '../../pages/Login';
 
 
 export default function LoginButton (){
