+++ client/src/index.js
@@ -10,6 +10,10 @@ import {
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
+import Login from './pages/Login';
+import Signup from './pages/Signup';
+
 
 const client = new ApolloClient({
   uri: 'https://48p1r2roz4.sse.codesandbox.io',
@@ -18,7 +22,27 @@ const client = new ApolloClient({
 
 ReactDOM.render(
   <ApolloProvider client={client}>
-    <App />
+    <App>
+    <div className= "container">
+      <Routes>
+        <Route>
+        <Route 
+                path="/" 
+                element={<App />} 
+              />
+          <Route 
+                path="/login" 
+                element={<Login />} 
+              />
+                <Route 
+                path="/signup" 
+                element={<Signup />} 
+              />
+        </Route>
+      </Routes>
+    </div>
+     
+    </App>
   </ApolloProvider>,
   document.getElementById('root')
 );
@@ -27,3 +51,4 @@ ReactDOM.render(
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
+
